+++ packages/at_client/test/encryption_util_test.dart
@@ -1,4 +1,5 @@
 import 'package:at_client/src/util/encryption_util.dart';
+import 'package:at_commons/at_commons.dart';
 import 'package:crypton/crypton.dart';
 import 'package:test/test.dart';
 
@@ -9,7 +10,7 @@ void main() {
       expect(aesKey, isNotEmpty);
     });
 
-    test('value encrypt/decrypt using aes key', () {
+    test('legacy encrypt/decrypt with AES', () {
       var aesKey = EncryptionUtil.generateAESKey();
       var valueToEncrypt = 'alice@atsign.com';
       var encryptedValue = EncryptionUtil.encryptValue(valueToEncrypt, aesKey);
@@ -17,7 +18,36 @@ void main() {
       expect(decryptedValue, valueToEncrypt);
     });
 
-    test('aes key encrypt/decrypt test', () {
+    test('encrypt/decrypt with AES', () {
+      var aesKey = EncryptionUtil.generateAESKey();
+      var iv = EncryptionUtil.generateIV();
+      var valueToEncrypt = 'alice@atsign.com';
+      var encryptedValue =
+          EncryptionUtil.encryptValue(valueToEncrypt, aesKey, ivBase64: iv);
+      var decryptedValue =
+          EncryptionUtil.decryptValue(encryptedValue, aesKey, ivBase64: iv);
+      expect(decryptedValue, valueToEncrypt);
+
+      expect(
+          () => EncryptionUtil.decryptValue(encryptedValue, aesKey),
+          throwsA(predicate((e) =>
+              e is AtKeyException &&
+              e.message ==
+                  'Invalid argument(s): Invalid or corrupted pad block')));
+
+      for (int i = 0; i < 10; i++) {
+        var otherIV = EncryptionUtil.generateIV();
+        expect(
+            () => EncryptionUtil.decryptValue(encryptedValue, aesKey,
+                ivBase64: otherIV),
+            throwsA(predicate((e) =>
+                e is AtKeyException &&
+                e.message ==
+                    'Invalid argument(s): Invalid or corrupted pad block')));
+      }
+    });
+
+    test('RSA encrypt/decrypt test', () {
       var aesKey = EncryptionUtil.generateAESKey();
       var rsaKeyPair = RSAKeypair.fromRandom();
       var encryptedKey =