+++ packages/at_client/test/notification_service_test.dart
@@ -60,11 +60,14 @@ class MockSharedKeyEncryption extends Mock implements SharedKeyEncryption {}
 // Mock class with implementation to populate metadata on encrypting value
 class MockSharedKeyEncryptionImpl extends Mock implements SharedKeyEncryption {
   @override
-  Future encrypt(AtKey atKey, value) async {
+  Future encrypt(AtKey atKey, value,
+      {bool storeSharedKeyEncryptedWithData = true}) async {
     //Set encryptionMetadata to atKey metadata
-    atKey.metadata = Metadata()
-      ..sharedKeyEnc = 'sharedKeyEnc'
-      ..pubKeyCS = 'publicKeyCS';
+    atKey.metadata = Metadata();
+    if (storeSharedKeyEncryptedWithData) {
+      atKey.metadata!.sharedKeyEnc = 'sharedKeyEnc';
+      atKey.metadata!.pubKeyCS = 'publicKeyCS';
+    }
     return 'encryptedValue';
   }
 }