+++ packages/at_client/test/sync_service_test.dart
@@ -46,7 +46,7 @@ class MockAtClient extends Mock implements AtClient {
   }
 
   @override
-  AtClientPreference? getPreferences() {
+  AtClientPreference getPreferences() {
     return AtClientPreference();
   }
 }